# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NagarDrishti AI — GIS Map API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface HeatmapPoint {
  lat: number;
  lng: number;
  /** 0.0–1.0, intensity for heatmap rendering */
  weight: number;
}

export type ComplaintClusterTopSeverity = typeof ComplaintClusterTopSeverity[keyof typeof ComplaintClusterTopSeverity];


export const ComplaintClusterTopSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface ComplaintCluster {
  id: string;
  lat: number;
  lng: number;
  count: number;
  topCategory: string;
  topSeverity: ComplaintClusterTopSeverity;
  wardName: string;
  ids: string[];
}

export type WardStatusSeverityLevel = typeof WardStatusSeverityLevel[keyof typeof WardStatusSeverityLevel];


export const WardStatusSeverityLevel = {
  none: 'none',
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface WardStatus {
  wardId: string;
  wardName: string;
  complaintCount: number;
  criticalCount: number;
  severityLevel: WardStatusSeverityLevel;
  centerLat: number;
  centerLng: number;
  /** Array of [lat, lng] pairs forming the ward boundary polygon */
  boundaryCoords: number[][];
}

export type NearbyComplaintCategory = typeof NearbyComplaintCategory[keyof typeof NearbyComplaintCategory];


export const NearbyComplaintCategory = {
  road: 'road',
  water: 'water',
  garbage: 'garbage',
  drainage: 'drainage',
  electricity: 'electricity',
  other: 'other',
} as const;

export type NearbyComplaintSeverity = typeof NearbyComplaintSeverity[keyof typeof NearbyComplaintSeverity];


export const NearbyComplaintSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export type NearbyComplaintStatus = typeof NearbyComplaintStatus[keyof typeof NearbyComplaintStatus];


export const NearbyComplaintStatus = {
  open: 'open',
  assigned: 'assigned',
  in_progress: 'in_progress',
  resolved: 'resolved',
} as const;

export interface NearbyComplaint {
  id: string;
  lat: number;
  lng: number;
  category: NearbyComplaintCategory;
  severity: NearbyComplaintSeverity;
  status: NearbyComplaintStatus;
  description: string;
  wardName: string;
  reporterCount: number;
  reportedAt: string;
  distanceKm: number;
}

export type EmergencyServiceType = typeof EmergencyServiceType[keyof typeof EmergencyServiceType];


export const EmergencyServiceType = {
  police: 'police',
  hospital: 'hospital',
  fire_station: 'fire_station',
  municipal: 'municipal',
} as const;

export interface EmergencyService {
  id: string;
  name: string;
  type: EmergencyServiceType;
  lat: number;
  lng: number;
  phone: string;
  distanceKm: number;
}

export interface MapStats {
  totalOpen: number;
  totalCritical: number;
  totalResolved: number;
  activeWards: number;
  last24hCount: number;
}

export type ComplaintCategory = typeof ComplaintCategory[keyof typeof ComplaintCategory];


export const ComplaintCategory = {
  road: 'road',
  water: 'water',
  garbage: 'garbage',
  drainage: 'drainage',
  electricity: 'electricity',
  other: 'other',
} as const;

export type ComplaintSeverity = typeof ComplaintSeverity[keyof typeof ComplaintSeverity];


export const ComplaintSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export type ComplaintStatus = typeof ComplaintStatus[keyof typeof ComplaintStatus];


export const ComplaintStatus = {
  open: 'open',
  assigned: 'assigned',
  in_progress: 'in_progress',
  resolved: 'resolved',
} as const;

export interface Complaint {
  id: string;
  lat: number;
  lng: number;
  category: ComplaintCategory;
  severity: ComplaintSeverity;
  status: ComplaintStatus;
  description: string;
  wardName: string;
  wardId: string;
  reporterCount: number;
  reportedAt: string;
  /** @nullable */
  imageUrl?: string | null;
  /**
     * AI-detected category from image or voice
     * @nullable
     */
  aiCategory?: string | null;
}

export type ComplaintInputCategory = typeof ComplaintInputCategory[keyof typeof ComplaintInputCategory];


export const ComplaintInputCategory = {
  road: 'road',
  water: 'water',
  garbage: 'garbage',
  drainage: 'drainage',
  electricity: 'electricity',
  other: 'other',
} as const;

export interface ComplaintInput {
  lat: number;
  lng: number;
  category: ComplaintInputCategory;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
}

export type GetHeatmapParams = {
category?: GetHeatmapCategory;
/**
 * Time window in hours (24, 48, 168 for 7 days)
 */
hours?: number;
};

export type GetHeatmapCategory = typeof GetHeatmapCategory[keyof typeof GetHeatmapCategory];


export const GetHeatmapCategory = {
  road: 'road',
  water: 'water',
  garbage: 'garbage',
  drainage: 'drainage',
  electricity: 'electricity',
  other: 'other',
  all: 'all',
} as const;

export type GetClustersParams = {
swLat: number;
swLng: number;
neLat: number;
neLng: number;
category?: GetClustersCategory;
};

export type GetClustersCategory = typeof GetClustersCategory[keyof typeof GetClustersCategory];


export const GetClustersCategory = {
  road: 'road',
  water: 'water',
  garbage: 'garbage',
  drainage: 'drainage',
  electricity: 'electricity',
  other: 'other',
  all: 'all',
} as const;

export type GetNearbyComplaintsParams = {
lat: number;
lng: number;
radiusKm?: number;
};

export type GetEmergencyServicesParams = {
lat: number;
lng: number;
};

export type ListComplaintsParams = {
status?: ListComplaintsStatus;
category?: string;
};

export type ListComplaintsStatus = typeof ListComplaintsStatus[keyof typeof ListComplaintsStatus];


export const ListComplaintsStatus = {
  open: 'open',
  assigned: 'assigned',
  in_progress: 'in_progress',
  resolved: 'resolved',
  all: 'all',
} as const;