# lib/api-zod/src/generated/types/complaint.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NagarDrishti AI — GIS Map API
 * OpenAPI spec version: 0.1.0
 */
import type { ComplaintCategory } from './complaintCategory';
import type { ComplaintSeverity } from './complaintSeverity';
import type { ComplaintStatus } from './complaintStatus';

export interface Complaint {
  id: string;
  lat: number;
  lng: number;
  category: ComplaintCategory;
  severity: ComplaintSeverity;
  status: ComplaintStatus;
  description: string;
  wardName: string;
  wardId: string;
  reporterCount: number;
  reportedAt: Date;
  /** @nullable */
  imageUrl?: string | null;
  /**
     * AI-detected category from image or voice
     * @nullable
     */
  aiCategory?: string | null;
}