# lib/api-zod/src/generated/types/complaintCluster.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NagarDrishti AI — GIS Map API
 * OpenAPI spec version: 0.1.0
 */
import type { ComplaintClusterTopSeverity } from './complaintClusterTopSeverity';

export interface ComplaintCluster {
  id: string;
  lat: number;
  lng: number;
  count: number;
  topCategory: string;
  topSeverity: ComplaintClusterTopSeverity;
  wardName: string;
  ids: string[];
}