# lib/api-zod/src/generated/types/complaintInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NagarDrishti AI — GIS Map API
 * OpenAPI spec version: 0.1.0
 */
import type { ComplaintInputCategory } from './complaintInputCategory';

export interface ComplaintInput {
  lat: number;
  lng: number;
  category: ComplaintInputCategory;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
}