# lib/api-zod/src/generated/types/emergencyService.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NagarDrishti AI — GIS Map API
 * OpenAPI spec version: 0.1.0
 */
import type { EmergencyServiceType } from './emergencyServiceType';

export interface EmergencyService {
  id: string;
  name: string;
  type: EmergencyServiceType;
  lat: number;
  lng: number;
  phone: string;
  distanceKm: number;
}