# lib/api-zod/src/generated/types/heatmapPoint.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NagarDrishti AI — GIS Map API
 * OpenAPI spec version: 0.1.0
 */

export interface HeatmapPoint {
  lat: number;
  lng: number;
  /** 0.0–1.0, intensity for heatmap rendering */
  weight: number;
}