# lib/api-zod/src/generated/types/mapStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NagarDrishti AI — GIS Map API
 * OpenAPI spec version: 0.1.0
 */

export interface MapStats {
  totalOpen: number;
  totalCritical: number;
  totalResolved: number;
  activeWards: number;
  last24hCount: number;
}