# lib/api-zod/src/generated/types/wardStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NagarDrishti AI — GIS Map API
 * OpenAPI spec version: 0.1.0
 */
import type { WardStatusSeverityLevel } from './wardStatusSeverityLevel';

export interface WardStatus {
  wardId: string;
  wardName: string;
  complaintCount: number;
  criticalCount: number;
  severityLevel: WardStatusSeverityLevel;
  centerLat: number;
  centerLng: number;
  /** Array of [lat, lng] pairs forming the ward boundary polygon */
  boundaryCoords: number[][];
}